python/requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
distro==1.8.0
fastapi==0.104.1
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
gradio==4.7.1
gradio_client==0.7.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
idna==3.6
importlib-resources==6.1.1
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.26.2
openai==0.28.0
orjson==3.9.10
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
psycopg2-binary==2.9.9
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
safetensors==0.4.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
tokenizers==0.15.0
tomlkit==0.12.0
toolz==0.12.0
torch==2.1.1
torchaudio==2.1.1
tqdm==4.66.1
transformers==4.35.2
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
websockets==11.0.3
yarl==1.9.3
